refactor: extract projects router (PR-10, Phase 1)
